backend/requirements.txt
# AgriPulse Backend - Requirements
# FastAPI & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
geoalchemy2==0.14.3

# Celery & Redis
celery[redis]==5.3.6
redis==5.0.1

# AWS S3 / DigitalOcean Spaces
boto3==1.34.34

# Sentinel Hub
sentinelhub==3.10.2

# Geospatial / Raster
rasterio==1.3.9
numpy==1.26.3
shapely==2.0.2
pyproj==3.6.1
rio-cogeo==5.1.1

# ML / PyTorch (for S4A model)
torch==2.1.2
pytorch-lightning==2.1.3

# Config & Utils
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2
